requirements.txt
asgiref==3.11.1
cloudinary==1.44.2
Django==6.0.4
django-cors-headers==4.9.0
django-cloudinary-storage==0.3.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
django-anymail[sendgrid]==10.2
whitenoise==6.12.0
